# openapi.yaml
openapi: 3.0.3
info:
  title: Dropbox Media Server API
  description: API for shared media gallery, public file requests, and administration.
  version: 1.6.0
servers:
  - url: /
paths:
...
  /api/droppr/requests:
    post:
      summary: Create a new file request link
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                path: { type: string }
                expires_hours: { type: integer }
                password: { type: string }
      responses:
        "200":
          description: Request link created
  /api/droppr/requests/{hash}:
    get:
      summary: Get info about a file request
      parameters:
        - $ref: "#/components/parameters/hash"
      responses:
        "200":
          description: Request details
  /api/droppr/requests/{hash}/upload:
    post:
      summary: Upload a file to a request
      parameters:
        - $ref: "#/components/parameters/hash"
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
                password: { type: string }
      responses:
        "200":
          description: File uploaded

  # User Management
...
components:
  parameters:
    hash:
      name: hash
      in: path
      required: true
      schema: { type: string }
...
  securitySchemes:
    XAuth:
      type: apiKey
      in: header
      name: X-Auth
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...